Move icao components around, re-encode txt files to utf-8, increase max build size
