+++ frontend/src/components/icao/AircraftTypes.vue
