+++ frontend/src/data/ICAO_Airlines.txt
@@ -12,7 +12,7 @@
 ;                                                                                                                         ;
 ;=========================================================================================================================;
 ;                                                                                                                         ;
-;  ALL RIGHTS RESERVED © AeroNav Association. NAV-DATA © AEROSOFT. FILES MUST NOT BE DISTRIBUTED TO NON-VATSIM MEMBERS!   ;
+;  ALL RIGHTS RESERVED Â© AeroNav Association. NAV-DATA Â© AEROSOFT. FILES MUST NOT BE DISTRIBUTED TO NON-VATSIM MEMBERS!   ;
 ;                                                                                                                         ;
 ;=========================================================================================================================;
 ;  The Global Nav Generator (GNG) and all it sub-modules where created by the AeroNav Association.                        ;
@@ -1610,7 +1610,7 @@ DYM	DYNAMIC AIR, S.L.	DYNAMIC SKY	SPAIN
 DYN	AERODYNAMICS, INC. (WATERFORD, MI)	AERODYNAMICS	UNITED STATES
 DZD	YAZD AIRWAYS	YAZD AIRWAYS	IRAN (ISLAMIC REPUBLIC OF)
 DZR	MIDWEST AVIATION (OMAHA, NE)	DOZER	UNITED STATES
-EAA	ELIT´AVIA INTERNATIONAL SM S.R.L.	IMOLA	SAN MARINO
+EAA	ELITÂ´AVIA INTERNATIONAL SM S.R.L.	IMOLA	SAN MARINO
 EAB	EAGLE ATLANTIC AIRLINES LIMITED	GOLD EAGLE	GHANA
 EAC	EXECUTIVE AIR CHARTER (HONOLULU, HI)	EXECAIR	UNITED STATES
 EAD	EAA - ESCOLA DE AVIACAO AEROCONDOR, SA	AERO-ESCOLA	PORTUGAL
@@ -4717,7 +4717,7 @@ PLB	POLYNESIAN BLUE	POLYBLUE	NEW ZEALAND
 PLC	POLICE AVIATION SERVICES	SPECIAL	UNITED KINGDOM
 PLF	POLISH AIR FORCE	POLISH AIRFORCE	POLAND
 PLG	MBK-S, OPEN JOINT STOCK COMPANY	PILGRIM	RUSSIAN FEDERATION
-PLI	AEROPERú	AEROPERU	PERU
+PLI	AEROPERÃº	AEROPERU	PERU
 PLL	AIR PAL, ESCUELA DE PILOTOS	AIRPAL	SPAIN
 PLM	WAMOS AIR S.A.	PULLMAN	SPAIN
 PLN	PLANAR - SOCIEDADE DE AVIACAO CIVIL SARL	PLANAR	ANGOLA
@@ -5823,7 +5823,7 @@ SXM	ST.MAARTEN AIRWAYS	SAINT MAARTEN	NETHERLANDS
 SXN	SAXON AIR	SAXONAIR	UNITED KINGDOM
 SXP	SOUTH PACIFIC EXPRESS HOLDINGS LIMITED	SOUTHPAC EXPRESS	SAMOA
 SXR	SKY EXPRESS	SKYSTORM	RUSSIAN FEDERATION
-SXS	SUNEXPRESS (GüNE? EKSPRES HAVAC?L?K A.?.)	SUNEXPRESS	TURKEY
+SXS	SUNEXPRESS (GÃ¼NE? EKSPRES HAVAC?L?K A.?.)	SUNEXPRESS	TURKEY
 SXT	SERVICIOS DE TAXI AEREOS, S.A. DE.C.V.	SERTAXI	MEXICO
 SXY	SAFARI EXPRESS CARGO LIMITED	SAFARI EXPRESS	KENYA
 SXZ	GREAT CIRCLE AVIATION, LLC (SAN MARCOS, CA)	SOCALAIR	UNITED STATES
